refactor(desktop): split chat window orchestration
