+++ apps/desktop/src/components/features/agents/agent-chat/use-agent-chat-window.ts
@@ -2,6 +2,17 @@ import type { RefCallback, RefObject } from "react";
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from "react";
 import type { AgentChatWindowRow } from "./agent-chat-thread-windowing";
 import { CHAT_OVERSCAN, CHAT_SHIFT_SIZE, CHAT_WINDOW_SIZE } from "./agent-chat-thread-windowing";
+import {
+  CHAT_AUTO_SCROLL_ANIMATION_DURATION_MS,
+  CHAT_MAX_RENDERED_ROWS,
+  clampWindowRange,
+  createBottomAnchoredWindow,
+  EMPTY_WINDOW,
+  type WindowRange,
+} from "./agent-chat-window-shared";
+import { useAgentChatResizeSync } from "./use-agent-chat-resize-sync";
+import { useAgentChatScrollController } from "./use-agent-chat-scroll-controller";
+import { useAgentChatSentinelObservers } from "./use-agent-chat-sentinel-observers";
 
 type UseAgentChatWindowInput = {
   rows: AgentChatWindowRow[];
@@ -24,49 +35,6 @@ type UseAgentChatWindowResult = {
   scrollToTop: () => void;
 };
 
-type WindowRange = {
-  start: number;
-  end: number;
-};
-
-type PendingScrollRequest = {
-  target: "top" | "bottom";
-  behavior: ScrollBehavior;
-  suppressSentinels: boolean;
-  animationDurationMs?: number;
-};
-
-const CHAT_AUTO_SCROLL_ANIMATION_DURATION_MS = 500;
-const CHAT_SCROLL_EDGE_THRESHOLD_PX = 48;
-const CHAT_SENTINEL_ROOT_MARGIN_PX = 200;
-const CHAT_MAX_RENDERED_ROWS = CHAT_WINDOW_SIZE + CHAT_OVERSCAN * 2;
-const EMPTY_WINDOW: WindowRange = { start: 0, end: -1 };
-
-const clampWindowRange = (range: WindowRange, rowCount: number): WindowRange => {
-  if (rowCount <= 0) {
-    return EMPTY_WINDOW;
-  }
-
-  const maxIndex = rowCount - 1;
-  const end = Math.max(0, Math.min(range.end, maxIndex));
-  const start = Math.max(0, Math.min(range.start, end));
-  return { start, end };
-};
-
-const createBottomAnchoredWindow = (rowCount: number): WindowRange => {
-  if (rowCount <= 0) {
-    return EMPTY_WINDOW;
-  }
-
-  return clampWindowRange(
-    {
-      start: Math.max(0, rowCount - CHAT_WINDOW_SIZE - CHAT_OVERSCAN),
-      end: rowCount - 1,
-    },
-    rowCount,
-  );
-};
-
 export function useAgentChatWindow({
   rows,
   activeSessionId,
@@ -77,134 +45,34 @@ export function useAgentChatWindow({
   const rowCount = rows.length;
   const initialWindow = createBottomAnchoredWindow(rowCount);
   const [windowRange, setWindowRange] = useState<WindowRange>(() => initialWindow);
-  const [isNearBottom, setIsNearBottom] = useState(true);
-  const [isNearTop, setIsNearTop] = useState(initialWindow.start === 0);
-  const [isAutoFollowingToBottom, setIsAutoFollowingToBottom] = useState(false);
-  const topObserverRef = useRef<IntersectionObserver | null>(null);
-  const bottomObserverRef = useRef<IntersectionObserver | null>(null);
-  const isUpdatingRef = useRef(false);
-  const prevScrollHeightRef = useRef<number | null>(null);
-  const shouldCompensateScrollRef = useRef(false);
   const prevSessionIdRef = useRef<string | null>(null);
   const prevIsSessionViewLoadingRef = useRef(isSessionViewLoading);
   const prevRowCountRef = useRef(rowCount);
-  const isPinnedToBottomRef = useRef(true);
-  const pendingScrollRequestRef = useRef<PendingScrollRequest | null>(null);
-  const suppressSentinelsRef = useRef(false);
-  const sentinelUnlockFrameRef = useRef<number | null>(null);
-  const scrollAnimationFrameRef = useRef<number | null>(null);
-  const scrollAnimationCleanupRef = useRef<(() => void) | null>(null);
-  const isBottomAutoFollowAnimationRef = useRef(false);
-  const contentResizeFrameRef = useRef<number | null>(null);
-  const observedContentHeightRef = useRef<number | null>(null);
-  const containerResizeFrameRef = useRef<number | null>(null);
-  const observedContainerHeightRef = useRef<number | null>(null);
-
-  const cancelScrollAnimation = useCallback((skipStateUpdate = false) => {
-    if (scrollAnimationFrameRef.current !== null && typeof window !== "undefined") {
-      window.cancelAnimationFrame(scrollAnimationFrameRef.current);
-    }
-    scrollAnimationFrameRef.current = null;
-
-    scrollAnimationCleanupRef.current?.();
-    scrollAnimationCleanupRef.current = null;
-
-    const wasBottomAutoFollow = isBottomAutoFollowAnimationRef.current;
-    isBottomAutoFollowAnimationRef.current = false;
-    if (!skipStateUpdate && wasBottomAutoFollow) {
-      setIsAutoFollowingToBottom(false);
-    }
-  }, []);
-
-  const animateScrollTo = useCallback(
-    (
-      container: HTMLDivElement,
-      resolveTargetTop: () => number,
-      durationMs: number,
-      trackBottomAutoFollow: boolean,
-    ) => {
-      cancelScrollAnimation();
-
-      const initialTargetTop = resolveTargetTop();
-      if (typeof window === "undefined") {
-        container.scrollTop = initialTargetTop;
-        return;
-      }
-
-      const startTop = container.scrollTop;
-      const initialDelta = initialTargetTop - startTop;
-      if (Math.abs(initialDelta) < 1 || durationMs <= 0) {
-        container.scrollTop = initialTargetTop;
-        return;
-      }
-
-      if (trackBottomAutoFollow) {
-        isBottomAutoFollowAnimationRef.current = true;
-        setIsAutoFollowingToBottom(true);
-      }
-
-      const handleUserInterrupt = () => {
-        cancelScrollAnimation();
-      };
-
-      container.addEventListener("wheel", handleUserInterrupt, { passive: true });
-      container.addEventListener("touchstart", handleUserInterrupt, { passive: true });
-      container.addEventListener("pointerdown", handleUserInterrupt, { passive: true });
-      scrollAnimationCleanupRef.current = () => {
-        container.removeEventListener("wheel", handleUserInterrupt);
-        container.removeEventListener("touchstart", handleUserInterrupt);
-        container.removeEventListener("pointerdown", handleUserInterrupt);
-      };
-
-      let startTime: number | null = null;
-      const step = (timestamp: number) => {
-        if (startTime === null) {
-          startTime = timestamp;
-        }
-
-        const elapsed = timestamp - startTime;
-        const progress = Math.min(1, elapsed / durationMs);
-        const currentTargetTop = resolveTargetTop();
-        container.scrollTop = startTop + (currentTargetTop - startTop) * progress;
-
-        if (progress >= 1) {
-          container.scrollTop = resolveTargetTop();
-          cancelScrollAnimation();
-          return;
-        }
-
-        scrollAnimationFrameRef.current = window.requestAnimationFrame(step);
-      };
-
-      scrollAnimationFrameRef.current = window.requestAnimationFrame(step);
-    },
-    [cancelScrollAnimation],
-  );
-
-  const requestWindowScroll = useCallback(
-    (request: PendingScrollRequest) => {
-      cancelScrollAnimation();
-      pendingScrollRequestRef.current = request;
-      if (!request.suppressSentinels) {
-        return;
-      }
-
-      suppressSentinelsRef.current = true;
-      if (sentinelUnlockFrameRef.current !== null && typeof window !== "undefined") {
-        window.cancelAnimationFrame(sentinelUnlockFrameRef.current);
-        sentinelUnlockFrameRef.current = null;
-      }
-    },
-    [cancelScrollAnimation],
-  );
+  const {
+    isNearBottom,
+    isNearTop,
+    isAutoFollowingToBottom,
+    isPinnedToBottomRef,
+    suppressSentinelsRef,
+    prevScrollHeightRef,
+    shouldCompensateScrollRef,
+    isUpdatingRef,
+    hasPendingScrollRequest,
+    syncBottomIfPinned,
+    requestWindowScroll,
+    applyPendingScrollRequest,
+    setBottomAnchoredState,
+    setTopAnchoredState,
+  } = useAgentChatScrollController({
+    messagesContainerRef,
+    initialWindow,
+  });
 
   const applyBottomAnchoredWindow = useCallback(() => {
     const nextWindow = createBottomAnchoredWindow(rowCount);
     setWindowRange(nextWindow);
-    setIsNearBottom(true);
-    setIsNearTop(nextWindow.start === 0);
-    isPinnedToBottomRef.current = true;
-  }, [rowCount]);
+    setBottomAnchoredState(nextWindow.start);
+  }, [rowCount, setBottomAnchoredState]);
 
   useEffect(() => {
     if (prevSessionIdRef.current === activeSessionId) {
@@ -241,9 +109,7 @@ export function useAgentChatWindow({
 
     if (rowCount === 0) {
       setWindowRange(EMPTY_WINDOW);
-      setIsNearBottom(true);
-      setIsNearTop(true);
-      isPinnedToBottomRef.current = true;
+      setBottomAnchoredState(0);
       return;
     }
 
@@ -257,7 +123,7 @@ export function useAgentChatWindow({
     }
 
     applyBottomAnchoredWindow();
-    if (pendingScrollRequestRef.current) {
+    if (hasPendingScrollRequest()) {
       return;
     }
 
@@ -267,216 +133,24 @@ export function useAgentChatWindow({
       suppressSentinels: false,
       animationDurationMs: CHAT_AUTO_SCROLL_ANIMATION_DURATION_MS,
     });
-  }, [applyBottomAnchoredWindow, requestWindowScroll, rowCount]);
+  }, [
+    applyBottomAnchoredWindow,
+    hasPendingScrollRequest,
+    isPinnedToBottomRef,
+    requestWindowScroll,
+    rowCount,
+    setBottomAnchoredState,
+  ]);
 
   useLayoutEffect(() => {
-    if (!shouldCompensateScrollRef.current) {
-      return;
-    }
-
-    shouldCompensateScrollRef.current = false;
-    const previousScrollHeight = prevScrollHeightRef.current;
-    prevScrollHeightRef.current = null;
-    const container = messagesContainerRef.current;
-    if (!container || previousScrollHeight === null) {
-      return;
-    }
-
-    const delta = container.scrollHeight - previousScrollHeight;
-    if (delta > 0) {
-      container.scrollTop += delta;
-    }
+    applyPendingScrollRequest();
   });
 
-  useLayoutEffect(() => {
-    const request = pendingScrollRequestRef.current;
-    if (!request) {
-      return;
-    }
-
-    pendingScrollRequestRef.current = null;
-    const container = messagesContainerRef.current;
-    if (!container) {
-      cancelScrollAnimation();
-      suppressSentinelsRef.current = false;
-      return;
-    }
-
-    if (typeof request.animationDurationMs === "number") {
-      animateScrollTo(
-        container,
-        () =>
-          request.target === "bottom"
-            ? Math.max(0, container.scrollHeight - container.clientHeight)
-            : 0,
-        request.animationDurationMs,
-        request.target === "bottom",
-      );
-    } else {
-      container.scrollTo({
-        top: request.target === "bottom" ? container.scrollHeight : 0,
-        behavior: request.behavior,
-      });
-    }
-
-    if (!request.suppressSentinels) {
-      return;
-    }
-
-    if (typeof window === "undefined") {
-      suppressSentinelsRef.current = false;
-      return;
-    }
-
-    sentinelUnlockFrameRef.current = window.requestAnimationFrame(() => {
-      suppressSentinelsRef.current = false;
-      sentinelUnlockFrameRef.current = null;
-    });
+  useAgentChatResizeSync({
+    messagesContainerRef,
+    messagesContentRef,
+    syncBottomIfPinned,
   });
-
-  const syncBottomIfPinned = useCallback(() => {
-    const container = messagesContainerRef.current;
-    if (!container) {
-      return;
-    }
-
-    if (pendingScrollRequestRef.current?.target === "bottom" || suppressSentinelsRef.current) {
-      return;
-    }
-
-    if (isBottomAutoFollowAnimationRef.current) {
-      setIsNearBottom(true);
-      return;
-    }
-
-    if (!isPinnedToBottomRef.current) {
-      return;
-    }
-
-    container.scrollTo({
-      top: container.scrollHeight,
-      behavior: "auto",
-    });
-    setIsNearBottom(true);
-    setIsNearTop(false);
-  }, [messagesContainerRef]);
-
-  useEffect(() => {
-    const container = messagesContainerRef.current;
-    const content = messagesContentRef.current;
-    if (!container || !content) {
-      observedContentHeightRef.current = null;
-      return;
-    }
-
-    observedContentHeightRef.current = content.scrollHeight;
-    if (typeof ResizeObserver === "undefined") {
-      return;
-    }
-
-    const syncAfterResize = () => {
-      contentResizeFrameRef.current = null;
-      const nextContent = messagesContentRef.current;
-      if (!nextContent) {
-        return;
-      }
-
-      const previousHeight = observedContentHeightRef.current;
-      const nextHeight = nextContent.scrollHeight;
-      observedContentHeightRef.current = nextHeight;
-      if (previousHeight === null || previousHeight === nextHeight) {
-        return;
-      }
-
-      syncBottomIfPinned();
-    };
-
-    const scheduleResizeSync = () => {
-      if (typeof window === "undefined") {
-        syncAfterResize();
-        return;
-      }
-
-      if (contentResizeFrameRef.current !== null) {
-        return;
-      }
-
-      contentResizeFrameRef.current = window.requestAnimationFrame(() => {
-        syncAfterResize();
-      });
-    };
-
-    const observer = new ResizeObserver(() => {
-      scheduleResizeSync();
-    });
-    observer.observe(content);
-
-    return () => {
-      observer.disconnect();
-      if (contentResizeFrameRef.current !== null && typeof window !== "undefined") {
-        window.cancelAnimationFrame(contentResizeFrameRef.current);
-        contentResizeFrameRef.current = null;
-      }
-    };
-  }, [messagesContainerRef, messagesContentRef, syncBottomIfPinned]);
-
-  useEffect(() => {
-    const container = messagesContainerRef.current;
-    if (!container) {
-      observedContainerHeightRef.current = null;
-      return;
-    }
-
-    observedContainerHeightRef.current = container.clientHeight;
-    if (typeof ResizeObserver === "undefined") {
-      return;
-    }
-
-    const syncAfterResize = () => {
-      containerResizeFrameRef.current = null;
-      const nextContainer = messagesContainerRef.current;
-      if (!nextContainer) {
-        return;
-      }
-
-      const previousHeight = observedContainerHeightRef.current;
-      const nextHeight = nextContainer.clientHeight;
-      observedContainerHeightRef.current = nextHeight;
-      if (previousHeight === null || previousHeight === nextHeight) {
-        return;
-      }
-
-      syncBottomIfPinned();
-    };
-
-    const scheduleResizeSync = () => {
-      if (typeof window === "undefined") {
-        syncAfterResize();
-        return;
-      }
-
-      if (containerResizeFrameRef.current !== null) {
-        return;
-      }
-
-      containerResizeFrameRef.current = window.requestAnimationFrame(() => {
-        syncAfterResize();
-      });
-    };
-
-    const observer = new ResizeObserver(() => {
-      scheduleResizeSync();
-    });
-    observer.observe(container);
-
-    return () => {
-      observer.disconnect();
-      if (containerResizeFrameRef.current !== null && typeof window !== "undefined") {
-        window.cancelAnimationFrame(containerResizeFrameRef.current);
-        containerResizeFrameRef.current = null;
-      }
-    };
-  }, [messagesContainerRef, syncBottomIfPinned]);
 
   const shiftWindowUp = useCallback(() => {
     if (isUpdatingRef.current || rowCount === 0 || suppressSentinelsRef.current) {
@@ -515,7 +189,14 @@ export function useAgentChatWindow({
       }
       return nextRange;
     });
-  }, [messagesContainerRef, rowCount]);
+  }, [
+    isUpdatingRef,
+    messagesContainerRef,
+    prevScrollHeightRef,
+    rowCount,
+    shouldCompensateScrollRef,
+    suppressSentinelsRef,
+  ]);
 
   const shiftWindowDown = useCallback(() => {
     if (isUpdatingRef.current || rowCount === 0 || suppressSentinelsRef.current) {
@@ -549,95 +230,22 @@ export function useAgentChatWindow({
       }
       return nextRange;
     });
-  }, [rowCount]);
+  }, [isUpdatingRef, rowCount, suppressSentinelsRef]);
 
-  const topSentinelRef = useCallback<RefCallback<HTMLDivElement>>(
-    (element) => {
-      topObserverRef.current?.disconnect();
-      topObserverRef.current = null;
-      if (!element || windowRange.start <= 0 || typeof IntersectionObserver === "undefined") {
-        return;
-      }
+  const { topSentinelRef, bottomSentinelRef } = useAgentChatSentinelObservers({
+    messagesContainerRef,
+    rowCount,
+    windowStart: windowRange.start,
+    windowEnd: windowRange.end,
+    suppressSentinelsRef,
+    shiftWindowUp,
+    shiftWindowDown,
+  });
 
-      const observer = new IntersectionObserver(
-        (entries) => {
-          const entry = entries[0];
-          if (entry?.isIntersecting) {
-            shiftWindowUp();
-          }
-        },
-        {
-          root: messagesContainerRef.current,
-          rootMargin: `${CHAT_SENTINEL_ROOT_MARGIN_PX}px 0px 0px 0px`,
-        },
-      );
-      observer.observe(element);
-      topObserverRef.current = observer;
-    },
-    [messagesContainerRef, shiftWindowUp, windowRange.start],
-  );
-
-  const bottomSentinelRef = useCallback<RefCallback<HTMLDivElement>>(
-    (element) => {
-      bottomObserverRef.current?.disconnect();
-      bottomObserverRef.current = null;
-      if (
-        !element ||
-        windowRange.end >= rowCount - 1 ||
-        typeof IntersectionObserver === "undefined"
-      ) {
-        return;
-      }
-
-      const observer = new IntersectionObserver(
-        (entries) => {
-          const entry = entries[0];
-          if (entry?.isIntersecting) {
-            shiftWindowDown();
-          }
-        },
-        {
-          root: messagesContainerRef.current,
-          rootMargin: `0px 0px ${CHAT_SENTINEL_ROOT_MARGIN_PX}px 0px`,
-        },
-      );
-      observer.observe(element);
-      bottomObserverRef.current = observer;
-    },
-    [messagesContainerRef, rowCount, shiftWindowDown, windowRange.end],
-  );
-
-  useEffect(() => {
-    const container = messagesContainerRef.current;
-    if (!container) {
-      return;
-    }
-
-    const handleScroll = () => {
-      const nearBottom =
-        container.scrollHeight - container.scrollTop - container.clientHeight <=
-        CHAT_SCROLL_EDGE_THRESHOLD_PX;
-      const nearTop = container.scrollTop <= CHAT_SCROLL_EDGE_THRESHOLD_PX;
-
-      const isEffectivelyPinned = isBottomAutoFollowAnimationRef.current || nearBottom;
-
-      setIsNearBottom(isEffectivelyPinned);
-      setIsNearTop(nearTop && windowRange.start === 0);
-      isPinnedToBottomRef.current = isEffectivelyPinned;
-    };
-
-    handleScroll();
-    container.addEventListener("scroll", handleScroll, { passive: true });
-    return () => {
-      container.removeEventListener("scroll", handleScroll);
-    };
-  }, [messagesContainerRef, windowRange.start]);
+  const effectiveIsNearTop = isNearTop && windowRange.start === 0;
 
   const scrollToBottom = useCallback(() => {
     applyBottomAnchoredWindow();
-    setIsNearBottom(true);
-    setIsNearTop(false);
-    isPinnedToBottomRef.current = true;
     requestWindowScroll({
       target: "bottom",
       behavior: "auto",
@@ -654,32 +262,13 @@ export function useAgentChatWindow({
       rowCount,
     );
     setWindowRange(nextRange);
-    isPinnedToBottomRef.current = false;
-    setIsNearBottom(false);
-    setIsNearTop(true);
+    setTopAnchoredState();
     requestWindowScroll({
       target: "top",
       behavior: "auto",
       suppressSentinels: true,
     });
-  }, [requestWindowScroll, rowCount]);
-
-  useEffect(() => {
-    return () => {
-      if (sentinelUnlockFrameRef.current !== null && typeof window !== "undefined") {
-        window.cancelAnimationFrame(sentinelUnlockFrameRef.current);
-      }
-      if (contentResizeFrameRef.current !== null && typeof window !== "undefined") {
-        window.cancelAnimationFrame(contentResizeFrameRef.current);
-      }
-      if (containerResizeFrameRef.current !== null && typeof window !== "undefined") {
-        window.cancelAnimationFrame(containerResizeFrameRef.current);
-      }
-      cancelScrollAnimation(true);
-      topObserverRef.current?.disconnect();
-      bottomObserverRef.current?.disconnect();
-    };
-  }, [cancelScrollAnimation]);
+  }, [requestWindowScroll, rowCount, setTopAnchoredState]);
 
   const effectiveWindow = clampWindowRange(windowRange, rowCount);
   const windowedRows =
@@ -692,7 +281,7 @@ export function useAgentChatWindow({
     windowStart: effectiveWindow.start,
     windowEnd: effectiveWindow.end,
     isNearBottom,
-    isNearTop,
+    isNearTop: effectiveIsNearTop,
     isAutoFollowingToBottom,
     topSentinelRef,
     bottomSentinelRef,
